Make transactions selection feature less biased.
 A node is free to have a strategy to choose specific transactions
and how many transactions to be added to the new block. However, the strategy must ensure that the selection
process is not biased to a specific node. It means that for example a node cannot wait only for its own
transactions or always ignore the transactions with zero transaction fee or lower fee. It can choose the highest
fee first, but if there are some other transactions in the pool, they need to be finally included in a block